+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 1.0.1 - 2020-01-15
+
+### Changed
+- restructure project to use tyarn workspaces: because of the way gatsby parses the file, the `npm link` feature
+  was not available to test the example site along with the current version of the plugin
+
 ## 1.0.0 - 2020-01-14
 
 ### Fixed